backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic-settings==2.7.0
pydantic[email]
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
httpx==0.27.0
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
anthropic==0.86.0
websockets==12.0
openpyxl==3.1.5
